@@ -1,12 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(<React.StrictMode>
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(<React.StrictMode>
   <App />
-</React.StrictMode>, document.getElementById('root'));
+</React.StrictMode>);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
